Split patron reference live tests
