+++ tests/Clc.Polaris.Api.LiveIntegrationTests/ReadOnly/Patron/PatronLanguagesGet.LiveTests.cs
@@ -1,0 +1,16 @@
+namespace Clc.Polaris.Api.LiveIntegrationTests
+{
+    [TestClass]
+    public sealed class PatronLanguagesGetTests : IntegrationTestBase
+    {
+        [TestMethod]
+        [ReadOnlyLiveTest]
+        public async Task PatronLanguagesGetAsync_ReturnsRows()
+        {
+            var response = await Papi.PatronLanguagesGetAsync(cancellationToken: TestContext.CancellationToken);
+
+            Assert.AreEqual(0, response.Data.PAPIErrorCode);
+            Assert.IsNotEmpty(response.Data.PatronLanguagesRows);
+        }
+    }
+}
